model FreezingDAO
